I18N;
Made ability for user to define custom field name. In this case if name is 'built-in' such field is added as TNT4J Event property.
Value (MsgValue) removed from 'built-in' fields as it was used as TNT4J event property.

Activity field lookup by name in 'built-in' fields enum StreamFieldType now is performed ignoring case.